Task Author Controller

+++ patika_dotnet_pratikler/BookStore/WebApiEf/Application/AuthorOperations/Command/Delete/DeleteAuthor.cs
@@ -1,3 +1,4 @@
+using Microsoft.EntityFrameworkCore;
 using WebApiEf.DbOperations;
 
 namespace WebApiEf.Application.AuthorOperations.Command.Delete
@@ -14,15 +15,20 @@ namespace WebApiEf.Application.AuthorOperations.Command.Delete
 
         public void Handle()
         {
-            var command = _context.Author.SingleOrDefault(x=> x.Id == Id);
+            var command = _context.Author.SingleOrDefault(x => x.Id == Id);
             if (command is null)
                 throw new InvalidOperationException("Yazar bulunamadı.");
-            
+
+            var checkAuthor = _context.Books.SingleOrDefault(x=> x.AuthorId == Id);
+
+            if (checkAuthor is not null)
+                throw new InvalidOperationException("Yazarın kitabı yayında önce yayında olan kitabı siliniz.");
+
             _context.Author.Remove(command);
             _context.SaveChanges();
         }
 
     }
 
-    
+
 }
